+++ 07-multi-db/src/db/strategies/mongodb.js
@@ -1,0 +1,13 @@
+const ICrud = require('./interfaces/interfaceCrud')
+
+class MongoDB extends ICrud{
+    constructor(){
+        super()
+    }
+
+    create(item){
+        console.log('O item foi salvo em MongoDB')
+    }
+}
+
+module.exports = MongoDB
