+++ 07-multi-db/src/db/strategies/postgres.js
@@ -1,0 +1,13 @@
+const ICrud = require('./interfaces/interfaceCrud')
+
+class Postgres extends ICrud{
+    constructor(){
+        super()
+    }
+
+    create(item){
+        console.log('O item foi salvo em Postgres')
+    }
+}
+
+module.exports = Postgres
