+++ 07-multi-db/src/index.js
@@ -1,0 +1,9 @@
+const ContextStrategy = require('./db/strategies/base/contextStrategy')
+const MongoDB = require('./db/strategies/mongodb')
+const Postgres = require('./db/strategies/postgres')
+
+const contextMongo = new ContextStrategy(new MongoDB())
+contextMongo.create()
+
+const contextPostgres = new ContextStrategy(new Postgres())
+contextPostgres.create()
